fixes-jutsu
just small fixes and added the possibility to eval only the experiments.

+++ experiment.py
@@ -9,6 +9,7 @@ import matplotlib.pyplot as plt
 import tensorflow as tf
 from sklearn.utils.class_weight import compute_class_weight
 from keras import backend as K
+from keras import mixed_precision
 from keras.callbacks import TensorBoard, BackupAndRestore, ModelCheckpoint, EarlyStopping, ReduceLROnPlateau
 from dataset import HDF5Dataset, TFDataset, load_ds_metadata, split_by_patients, split_by_centers, rus, trim_sets
 from utilities import save_split_data, load_split_data, plot_charts
@@ -24,6 +25,7 @@ class Experiment:
                  ds_map_pkl:str,
                  ds_split_pkl:str,
                  results_dir:str,
+                 eval_only:bool = False,
                  workers:int = 1,
                  shuffle_buffer_size:int = 100,
                  max_queue_size:int = 512,
@@ -38,6 +40,7 @@ class Experiment:
         self.ds_map_pkl = ds_map_pkl
         self.ds_split_pkl = ds_split_pkl
         self.results_dir = results_dir
+        self.eval_only = eval_only
         self.shuffle_buffer_size = shuffle_buffer_size
         self.max_queue_size = max_queue_size
         self.workers = workers
@@ -46,7 +49,7 @@ class Experiment:
         self.seed = random_state
 
         # params to be computed
-        self.hw_accel = False
+        self.device = None
         self.settings = None
         self.exp_name = ''
         self.exp_results_subdir = ''
@@ -87,8 +90,13 @@ class Experiment:
     def build(self):
         # setting os environment
         os.environ['TF_CPP_MIN_LOG_LEVEL'] = '1'
+        
+        #TODO: da attivare
+        #tf.config.optimizer.set_jit(True)
+        #policy = mixed_precision.Policy("mixed_float16")
+        #mixed_precision.set_global_policy(policy)
 
-        self.check_hw_accel()
+        self.is_gpu_available()
         self.load_dataset()
         self.init_results()
     
@@ -98,20 +106,16 @@ class Experiment:
             return json.load(file)
 
 
-    def check_hw_accel(self):
+    def is_gpu_available(self):
+        cpu_device = tf.config.list_physical_devices('CPU')[0]
         gpu_devices = tf.config.list_physical_devices('GPU')
-
+        
         if gpu_devices:
-            # add CPU 0 and GPU 0 to the available devices
-            cpu_device = tf.config.list_physical_devices('CPU')[0]
-            gpu_device = gpu_devices[0]
-            
-            tf.config.set_visible_devices([gpu_device, cpu_device])
-            tf.config.experimental.set_memory_growth(gpu_device, True)
+            # "/GPU:0"
+            self.device = gpu_devices[0]
+        else:
+            self.device = cpu_device
 
-            # set the class attribute regarding the hardware acceleration 
-            self.hw_accel = True
-    
 
     def load_exp_settings(self, exp_idx):
         # load the requested experiment settings by using the index to find it in the json file
@@ -119,10 +123,14 @@ class Experiment:
         
         # build the experiment name based on the configuration extracted
         self.exp_name = self.build_exp_name()
-
+        
         # create the experiment results subdirectory
         self.exp_results_subdir = os.path.join(self.results_dir, self.exp_name)
         os.makedirs(self.exp_results_subdir, exist_ok=True)
+
+        # create the weights subdirectory
+        weights_path = os.path.join(self.exp_results_subdir, 'weights/') 
+        os.makedirs(weights_path, exist_ok=True)
 
         # reset previous configurations
         self.metrics_results = {}
@@ -133,7 +141,8 @@ class Experiment:
 
     def build_exp_name(self):
         # parameters not to be used to generate the experiment name
-        excl_params = ["ds_split_ratio", "ds_trim", "metrics"]
+        # excl_params = ["ds_split_ratio", "ds_trim", "metrics"]
+        excl_params = ["ds_split_ratio", "metrics"]
         experiment_params = {key: value for key, value in self.settings.items() if key not in excl_params}
 
         # generate the experiment name based on the parameters
@@ -152,22 +161,33 @@ class Experiment:
         self.csv_results_path = os.path.join(self.results_dir, "results.csv")
         self.csv_columns = ["experiment", "ccr", "mae", "ms", "rmse", "acc_1off", "qwk"]
         
+        # ask for a confirmation
+        confirmation = input("Do you want to delete existing run files? (y/[N]): ").lower()
+        
         # experiments checkpoint file 
-        if not os.path.exists('run_checkpoint.txt'):
+        if not os.path.exists('run_checkpoint.txt') and confirmation == 'y':
             # clean the previous results and re-make the directory
             if os.path.exists(self.results_dir):
                 shutil.rmtree(self.results_dir)
             os.makedirs(self.results_dir)
+            
+            # remove previous logs/ directory
+            if os.path.exists('./logs/'):
+                shutil.rmtree('./logs/')
 
-            # remove previous logs/ and weights/ directories
-            for directory in ['./logs/', './weights/']:
-                if os.path.exists(directory):
-                    shutil.rmtree(directory)
+            # delete splitting file
+            if os.path.exists("./splitdata.pkl"):
+                os.remove("./splitdata.pkl")
 
             # create the CSV file and write the header
             with open(self.csv_results_path, mode='w', encoding='UTF-8', newline='') as csv_file:
                 writer = csv.writer(csv_file)
                 writer.writerow(self.csv_columns)
+        
+        if self.eval_only:
+            # delete splitting file
+            if os.path.exists("./splitdata.pkl"):
+                os.remove("./splitdata.pkl")
 
     
     def check_exps_ckp(self, curr_exp):
@@ -202,7 +222,7 @@ class Experiment:
         
         # trim the dataset if requested
         if ds_trim > 0:
-            trimmed_sets = trim_sets(self.idxs_train, self.idxs_val, self.idxs_test, ds_trim)
+            trimmed_sets = trim_sets(self.idxs_train, self.idxs_test, self.idxs_val, ds_trim, self.seed)
             self.idxs_train, self.idxs_val, self.idxs_test = trimmed_sets
         
         # extract the train, val and test set labels
@@ -210,15 +230,15 @@ class Experiment:
         self.y_val = self.dataset_labels[self.idxs_val]
         self.y_test = self.dataset_labels[self.idxs_test]
         
-        # export the splitting if shared between experiments
-        if exps_common_settings:
-            split_data = {'train': {'x': self.idxs_train, 'y': self.y_train}, 
-                          'val': {'x': self.idxs_val, 'y': self.y_val}, 
-                          'test': {'x': self.idxs_test, 'y': self.y_test}, 
-                          'metadata': self.dataset_metadata}
-            save_split_data(split_data, self.results_dir)
-    
+        # export the splitting
+        split_data = {'train': {'x': self.idxs_train, 'y': self.y_train}, 
+                        'val': {'x': self.idxs_val, 'y': self.y_val}, 
+                        'test': {'x': self.idxs_test, 'y': self.y_test}, 
+                        'metadata': self.dataset_metadata}
+        save_path = self.results_dir if exps_common_settings else self.exp_results_subdir
+        save_split_data(split_data, save_path)
 
+    
     def load_dataset_splitted(self):
         ltrain, lval, ltest, lmetadata = load_split_data(self.results_dir)
         self.idxs_train, self.y_train = ltrain['x'], ltrain['y']
@@ -237,10 +257,11 @@ class Experiment:
                                  self.idxs_train, 
                                  batch_size=batch_size, 
                                  buffer_size=self.shuffle_buffer_size, 
-                                 is_train=True, 
-                                 augmentation=augmentation).as_iterator()
-        self.x_val = TFDataset(self.dataset, self.idxs_val, batch_size=batch_size).as_iterator()
-        self.x_test = TFDataset(self.dataset, self.idxs_test, batch_size=batch_size).as_iterator()
+                                 is_train=True,
+                                 augmentation=augmentation,
+                                 device=self.device).as_iterator()
+        self.x_val = TFDataset(self.dataset, self.idxs_val, batch_size=batch_size, device=self.device).as_iterator()
+        self.x_test = TFDataset(self.dataset, self.idxs_test, batch_size=batch_size, device=self.device).as_iterator()
         
         
     def compute_class_weight(self):
@@ -248,7 +269,7 @@ class Experiment:
         train_class_weights = compute_class_weight('balanced', classes=np.unique(self.y_train), y=self.y_train)
         train_class_weights = np.round(train_class_weights, 4)
         self.train_class_weights = dict(enumerate(train_class_weights))
-        
+        return self.train_class_weights
     
     def generate_split_charts(self, charts=None, per_exp=False):
         if self.dataset_metadata is not None:
@@ -326,7 +347,7 @@ class Experiment:
             loss = qwk_loss(cost_matrix)
         
         # metrics
-        metrics_t = Metrics(self.ds_num_classes, self.settings['nn_type'], 'train')
+        metrics_t = Metrics(self.ds_num_classes, self.settings['nn_type'])
         train_metrics = [getattr(metrics_t, metric_name) for metric_name in metrics if metric_name not in self.train_metrics_exl]
 
         # compile
@@ -341,11 +362,12 @@ class Experiment:
         # parameters
         epochs = self.settings['nn_epochs']
         batch_size = self.settings['nn_batch_size']
-        
+        ckpt_filename = os.path.join(self.exp_results_subdir, "weights", "best_ckpt.hdf5")
+
         # callbacks
         tensorboard = TensorBoard(log_dir=f"logs/fit/{self.exp_name}", histogram_freq=1)
         backup = BackupAndRestore(backup_dir="backup/")
-        checkpoint = ModelCheckpoint(f'weights/{self.exp_name}', monitor='val_loss', save_weights_only=True, save_best_only=True, verbose=self.verbose)
+        checkpoint = ModelCheckpoint(ckpt_filename, monitor='val_loss', save_weights_only=True, save_best_only=True, verbose=self.verbose)
         early_stop = EarlyStopping(monitor='val_loss', patience=10, verbose=self.verbose)
         reduce_lr = ReduceLROnPlateau(monitor='val_loss', factor=0.2, patience=6, min_lr=1e-6, verbose=self.verbose)
         gradcam = GradCAMCallback(model, self, freq=gradcam_freq) if gradcam_freq > 0 else None
@@ -417,29 +439,23 @@ class Experiment:
         plt.close()
 
 
-    def nn_model_evaluate(self, model, load_best_weights=True):
-        # Load the best weights
-        if load_best_weights:
-            model.load_weights(f'weights/{self.exp_name}')
-
+    def nn_model_evaluate(self, model, weights=None, load_best_weights=True):
+        # if specified, load the weights passed as argument (priority)
+        if weights is not None:
+            model.load_weights(weights)
+        elif load_best_weights:
+            # load the best weights
+            best_weights_file = os.path.join(self.exp_results_subdir, "weights", "best_ckpt.hdf5")
+            try:
+                model.load_weights(best_weights_file)
+            except Exception as e:
+                raise Exception('error while loading best weights file: ', e)
+        
         # get the batch size
         nn_batch_size = self.settings['nn_batch_size']
-
-        # steps
-        #test_samples = len(self.idxs_test)
-        #test_steps = -(-test_samples // nn_batch_size)  # Equivalent to ceil(test_samples / nn_batch_size)
-
-        # TODO: check if evaluate give same results as manual metrics computing
-        # model.evaluate(self.x_test, 
-        #                steps=len(self.idxs_test) // nn_batch_size,
-        #                max_queue_size=self.max_queue_size,
-        #                workers=self.workers,
-        #                use_multiprocessing=False,
-        #                verbose=self.verbose)
         
         # model evaluation, get the predictions by running the model inference
         y_test_pred = model.predict(self.x_test, 
-                                    #steps=len(self.idxs_test) // nn_batch_size,
                                     steps=-(-len(self.idxs_test) // nn_batch_size),
                                     verbose=self.verbose,
                                     max_queue_size=self.max_queue_size,
@@ -448,7 +464,7 @@ class Experiment:
                                     )
         
         # compute evaluation metrics
-        metrics_e = Metrics(self.ds_num_classes, self.settings['nn_type'], 'eval')
+        metrics_e = Metrics(self.ds_num_classes, self.settings['nn_type'])
         eval_metrics = [(getattr(metrics_e, metric_name), metric_name) for metric_name in self.settings['metrics']]
         
         for metric, metric_name  in eval_metrics:
@@ -469,7 +485,7 @@ class Experiment:
         with open(self.csv_results_path, mode='a', encoding='UTF-8', newline='') as csv_file:
             writer = csv.writer(csv_file)
             writer.writerow(values_columns)
-
+        
         # update the experiments checkpoint file 
         datetime_log = datetime.now().strftime("%m/%d/%Y,%H:%M:%S")
         with open('run_checkpoint.txt', mode='a', encoding='UTF-8') as ckp_file:
